+++ app/src/main/java/com/tovo/eat/ui/account/orderhistory/historylist/OrdersHistoryListResponse.java
@@ -45,10 +45,10 @@ public class OrdersHistoryListResponse {
 
         @SerializedName("orderid")
         @Expose
-        public Integer orderid;
+        public Long orderid;
         @SerializedName("userid")
         @Expose
-        public Integer userid;
+        public Long userid;
         @SerializedName("ordertime")
         @Expose
         public String ordertime;
@@ -140,19 +140,19 @@ public class OrdersHistoryListResponse {
         @Expose
         public List<Item> items = null;
 
-        public Integer getOrderid() {
+        public Long getOrderid() {
             return orderid;
         }
 
-        public void setOrderid(Integer orderid) {
+        public void setOrderid(Long orderid) {
             this.orderid = orderid;
         }
 
-        public Integer getUserid() {
+        public Long getUserid() {
             return userid;
         }
 
-        public void setUserid(Integer userid) {
+        public void setUserid(Long userid) {
             this.userid = userid;
         }
