+++ app/src/main/java/com/tovo/eat/ui/address/select/SelectAddressListViewModel.java
@@ -53,7 +53,7 @@ public class SelectAddressListViewModel extends BaseViewModel<SelectAddressListN
         getNavigator().addNewAddress();
     }
 
-    public void defaultAddress(Integer aid) {
+    public void defaultAddress(Long aid) {
 
         try {
             setIsLoading(true);
@@ -81,7 +81,7 @@ public class SelectAddressListViewModel extends BaseViewModel<SelectAddressListN
     }
 
 
-    public void updateCurrentAddress(String title, String address, double lat, double lng, String area, Integer aid) {
+    public void updateCurrentAddress(String title, String address, double lat, double lng, String area, Long aid) {
 
         getDataManager().updateCurrentAddress(title, address, lat, lng, area, aid);
