+++ app/src/main/java/com/tovo/eat/ui/cart/CartNavigator.java
@@ -36,5 +36,5 @@ public interface CartNavigator {
 
     void notServicable();
     void funnelAlert();
-    void gotoKitchen(int kitchenid);
+    void gotoKitchen(Long kitchenid);
 }
